sync: Import upstream OpenBSD 4.9
Import OpenBSD 4.9 and make the current L4 port reflect the changes.

Important changes affecting the L4 port include:

 - ci_cpuid converts from 32 bits to uint64_t
 - netisrs are handled as generic softintrs at IPL_SOFTNET
 - evcount_intr for evcount_attach() is gone for good
 - PCIe support (which is not ported yet)

For a more complete list of changes, visit:

  http://www.openbsd.org/plus49.html

+++ usr.sbin/cron/crontab.c
@@ -1,4 +1,4 @@
-/*	$OpenBSD: crontab.c,v 1.58 2009/10/27 23:52:16 deraadt Exp $	*/
+/*	$OpenBSD: crontab.c,v 1.60 2011/02/11 07:14:49 guenther Exp $	*/
 
 /* Copyright 1988,1990,1993,1994 by Paul Vixie
  * All rights reserved
@@ -56,9 +56,9 @@ static	void		list_cmd(void),
 			edit_cmd(void),
 			check_error(const char *),
 			parse_args(int c, char *v[]),
+			copy_crontab(FILE *, FILE *),
 			die(int);
 static	int		replace_cmd(void);
-static	int		ignore_comments(FILE *);
 
 static void
 usage(const char *msg) {
@@ -251,12 +251,7 @@ list_cmd(void) {
 	 */
 	Set_LineNum(1)
 
-	/* ignore the top few comments since we probably put them there.
-	 */
-	ch = ignore_comments(f);
-
-	while (EOF != (ch = get_char(f)))
-		putchar(ch);
+	copy_crontab(f, stdout);
 	fclose(f);
 }
 
@@ -288,6 +283,7 @@ check_error(const char *msg) {
 static void
 edit_cmd(void) {
 	char n[MAX_FNAME], q[MAX_TEMPSTR];
+	const char *tmpdir;
 	FILE *f;
 	int ch, t;
 	struct stat statbuf, xstatbuf;
@@ -329,12 +325,26 @@ edit_cmd(void) {
 	(void)signal(SIGINT, SIG_IGN);
 	(void)signal(SIGQUIT, SIG_IGN);
 
-	if (snprintf(Filename, sizeof Filename, "%scrontab.XXXXXXXXXX",
-	    _PATH_TMP) >= sizeof(Filename)) {
+	tmpdir = getenv("TMPDIR");
+	if (tmpdir == NULL || tmpdir[0] == '\0')
+		tmpdir = _PATH_TMP;
+	for (t = strlen(tmpdir); t != 0 && tmpdir[t - 1] == '/'; t--)
+		continue;
+	if (snprintf(Filename, sizeof Filename, "%.*s/crontab.XXXXXXXXXX",
+	    t, tmpdir) >= sizeof(Filename)) {
 		fprintf(stderr, "path too long\n");
 		goto fatal;
 	}
-	if (-1 == (t = mkstemp(Filename))) {
+	if (swap_gids() < OK) {
+		perror("swapping gids");
+		exit(ERROR_EXIT);
+	}
+	t = mkstemp(Filename);
+	if (swap_gids_back() < OK) {
+		perror("swapping gids back");
+		exit(ERROR_EXIT);
+	}
+	if (t == -1) {
 		perror(Filename);
 		goto fatal;
 	}
@@ -345,15 +355,7 @@ edit_cmd(void) {
 
 	Set_LineNum(1)
 
-	/* ignore the top few comments since we probably put them there.
-	 */
-	ch = ignore_comments(f);
-
-	/* copy the rest of the crontab (if any) to the temp file.
-	 */
-	if (EOF != ch)
-		while (EOF != (ch = get_char(f)))
-			putc(ch, NewCrontab);
+	copy_crontab(f, NewCrontab);
 	fclose(f);
 	if (fflush(NewCrontab) < OK) {
 		perror(Filename);
@@ -366,7 +368,15 @@ edit_cmd(void) {
 		fprintf(stderr, "%s: error while writing new crontab to %s\n",
 			ProgramName, Filename);
  fatal:
+		if (swap_gids() < OK) {
+			perror("swapping gids");
+			exit(ERROR_EXIT);
+		}
 		unlink(Filename);
+		if (swap_gids_back() < OK) {
+			perror("swapping gids back");
+			exit(ERROR_EXIT);
+		}
 		exit(ERROR_EXIT);
 	}
 
@@ -387,10 +397,18 @@ edit_cmd(void) {
 		goto fatal;
 	}
 	if (timespeccmp(&mtimespec, &statbuf.st_mtimespec, -) == 0) {
+		if (swap_gids() < OK) {
+			perror("swapping gids");
+			exit(ERROR_EXIT);
+		}
 		if (lstat(Filename, &xstatbuf) == 0 &&
 		    statbuf.st_ino != xstatbuf.st_ino) {
 			fprintf(stderr, "%s: crontab temp file moved, editor "
 			   "may create backup files improperly\n", ProgramName);
+		}
+		if (swap_gids_back() < OK) {
+			perror("swapping gids back");
+			exit(ERROR_EXIT);
 		}
 		fprintf(stderr, "%s: no changes made to crontab\n",
 			ProgramName);
@@ -432,7 +450,15 @@ edit_cmd(void) {
 		goto fatal;
 	}
  remove:
+	if (swap_gids() < OK) {
+		perror("swapping gids");
+		exit(ERROR_EXIT);
+	}
 	unlink(Filename);
+	if (swap_gids_back() < OK) {
+		perror("swapping gids back");
+		exit(ERROR_EXIT);
+	}
  done:
 	log_it(RealUser, Pid, "END EDIT", User);
 }
@@ -647,14 +673,16 @@ die(int x) {
 	_exit(ERROR_EXIT);
 }
 
-static int
-ignore_comments(FILE *f) {
+static void
+copy_crontab(FILE *f, FILE *out) {
 	int ch, x;
 
+	/* ignore the top few comments since we probably put them there.
+	 */
 	x = 0;
 	while (EOF != (ch = get_char(f))) {
 		if ('#' != ch) {
-			putc(ch, NewCrontab);
+			putc(ch, out);
 			break;
 		}
 		while (EOF != (ch = get_char(f)))
@@ -664,5 +692,9 @@ ignore_comments(FILE *f) {
 			break;
 	}
 
-	return ch;
+	/* copy out the rest of the crontab (if any)
+	 */
+	if (EOF != ch)
+		while (EOF != (ch = get_char(f)))
+			putc(ch, out);
 }
